Memories_chhoti/requirements.txt
Flask==3.0.0
flask-cors
google-genai
python-dotenv
